Move sample program to `code` textarea
The sample program was hard coded into the language interpreter since
that made initial testing easier. However, we need to be able to
dynamically load a program from the `code` textarea and spit the
results out to the `results` textarea in order for the web IDE to be
useful.

This commit wires the core interpreter pipeline up with the two
textareas so that arbitrary programs can be written and executed.