WIP Integreate parser/evaluator with IDE
There's a lot going on here so I'm going to try to explain what's in my
mind so far.

I copied the `Codeable_Eval_Parser.pl` from the `src` directory into
the `runtime` directory and renamed it to `codeable.pl`. I made some
code updates there to better align with the `grammar.md`

I tried to point the `ide.js` to the `codeable.pl` but it's failing to
load the code from it for some reason so I copied the code into
`jhaleAssign3.pl` (literally the only difference is the filename, so
I'm not sure why `ide.js` likes one but not the other).

But, the one success is the `show` keyword now works in the IDE!